Introduce $JUJU_HOME/ssh for client keys
We introduce a new ssh directory under $JUJU_HOME,
in which we create an passphrase-less RSA key pair
for use by an SSH client. The user may add additional
key pairs to the directory. All public keys in the
directory (*.pub with a corresponding file without
the suffix) are added to the authorized-keys
environment config.

The private keys in the directory will be used in a
later CL that introduces a go.crypto/ssh-based client,
which will be used when openssh is not available.