Introduce utils/ssh.Client interface
This is a refactoring of utils/ssh to introduce
a Client interface which consists of Command and
Copy methods. The existing Command and ScpCommand
methods have been subsumed by the new OpenSSHClient
implementation.

All of the packages using ssh.Command/ScpCommand
have been updated to use the new client interface.
The options have also be changed to a struct which
is interpreted by the implementation. There is one
change in default options: password authentication
is now disabled by default, and must be explicitly
enabled.

There is a new option, which is identity files to
specify when authenticating. Clients are expected
to give these precedence over their client-specific
defaults (~/.ssh/id_rsa, etc.)

Parts of environs/manual/fake.ssh have been moved
to utils/ssh/testing.