utils/ssh: explicitly specify default identities
There is a bug in ssh_openssh.go: if any identities
are specified, then default identities are not
attempted unless they are loaded into the ssh-agent.
This is only a problem if the user relies on the
default identity, AND explicitly specifies
authorized-keys in their environment config.

This problem is fixed by explicitly specifying the
default identities if any identities are specified,
which is almost always (~/.juju/ssh).

Fixes lp:1275657