Added proper symlink support to tar/untar
Symlinks where being treated as regular files and copies
of the files where stored instead of links for files and
broken regular files with no content instead of folders.
Now Lstat is being used to read the file details, which
does not follow symlinks but instead returns the link
details, which allows to add them as symlinks to the tar.