Fix LP 1536378
Fixes LP #1536378

The logic inside isAlive was wrong, it was never executing the wait loop
so if two independent processes called isAlive (via clean) then they would
both think the lock is stale and attempt to break it.

The number of loop iterations if the alive file cannot be read is set to
an arbitary value which will limit the max delay in tying to lock a dead lock
to 300 seconds.