Add dependencies.tsv
The master branch for juju/utils had a dependencies.tsv
file added some time after 1.25 was released.  juju/utils
needed to be forked to fix an issue for 1.25 without
pulling in all the changes commited since then.

The merge job has been changed to look for the dependencies
file in order to merge PRs into the repo.  I copied the
current dependencies.tsv file from the master branch and
verified that this 1.25 branch for juju/utils built and
passed go test with those dependencies.