1568312 - fail install after first "not found" err
We should stop trying to install packages if the
error indicates that the package could not be
found / does not exist.

Otherwise, we will spend an extra 5 minutes during
bootstrap waiting on a package that will never
install.

This 5 extra minutes causes Juju's CI tests to
timeout and is a bad user experience.