Address review comments
I moved the check for a fatal error to be specific to the
package manager.

I also moved the check for a fatal error to after the check
if the error code was != 100.  This saved adding in a special
case for yum since it would return 1 in the case that the
package(s) could not be found.