Make natural sorting implementation more generic
Multiple runs of digits are now supported by the natural sorting
algorihm. In Juju's case this means that machine ids for containers
are correctly sorted, as are unit names which contain digits.

Testing of the algorithm has been simplified. Instead of having to
provide an input and the desired output, only the desired output needs
to be provided. The test infrastructure will test by sorting a reverse
of the input as well as a randomly shuffled versions of the
input. This makes writing test cases much easier.

Test cases have also been made more orthogonal with less overlap. The
unnecessarily long test names have also been cleaned up.