retry: new package
The plan is that this should eventually replace utils.Attempt
and github.com/juju/try.

It provides functionality for stopping and clock mocking and for pluggable
strategies.

It is (arguably) quite a bit simpler to use than github.com/juju/try and
the usage remains basically compatible with utils.Attempt, so
only minor code changes will be needed to change the existing
code base to use it.