ssh: Add StripCRReader for gocrypto on windows
Fixes lp:1468752 by removing carriage returns from textual stdin
stream on windows when using gocrypto to communicate with remote
machine.

This is part one of making ssh from windows to a unix machine
more usable. Notably, this doesn't enable the passing of control
codes through to the remote shell, so for instance using a text
editor is still problematic.