ssh: host-key checking for crypto/ssh client
We introduce a host-key checker that mimics
the behaviour of the OpenSSH client. The
default "StrictHostKeyChecking" behaviour
is now for the crypto/ssh client to ask
(if a terminal is available, else fail),
and to leave off the option for openssh
to defer the decision to the openssh client
code.

We also make the host key algorithms
configurable. This will be used in a Juju
PR to limit which host key algorithms are
acceptable.