IsFileOwner is file owned by user
The following adds the ability to check if the current file is owned
by the user supplied. If the user can not be found, or no information
about the file can be found, then it will error out.

The windows version is not supplied, as there isn't a need for it,
but could be researched to see if it could be.